requirements.txt
pandas
numpy
pyyaml
requests
tqdm
